getCarreras movido a actions en comun
