+++ app/lib/data/actions.ts
@@ -264,3 +264,22 @@ export const obtenerDatosUsuarioFetch = async () => {
     return null;
   }
 };
+
+export const getCarreras = async () => {
+  const token = authToken();
+  if (token) {
+    const response = await fetch(`${apiRoute}/carreras`, {
+      method: 'GET',
+      headers: {
+        Authotization: `Bearer ${token}`,
+      },
+    });
+    if (response.ok) {
+      const data = await response.json();
+      console.log(data);
+      return data;
+    } else {
+      return null;
+    }
+  }
+};
