+++ app/lib/data/funcionario/actions.ts
@@ -187,25 +187,6 @@ export async function registrarPeriodoExamen(
   }
 }
 
-export const getCarreras = async () => {
-  const token = authToken();
-  if (token) {
-    const response = await fetch(`${apiRoute}/carreras`, {
-      method: 'GET',
-      headers: {
-        Authotization: `Bearer ${token}`,
-      },
-    });
-    if (response.ok) {
-      const data = await response.json();
-      console.log(data);
-      return data;
-    } else {
-      return null;
-    }
-  }
-};
-
 export async function getPeriodosExamenCarrera(id: string) {
   const token = authToken();
   const carreraJson = await getCarrera(id);
